[playground] Viteプラグインを使って、vueファイルのレンダリングができるようになった（styleは未対応）
