+++ examples/playground/vite.config.js
@@ -2,11 +2,15 @@ import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'
 
+import chibivue from '../../packages/@extensions/vite-plugin-chibivue'
+
 const dirname = path.dirname(fileURLToPath(new URL(import.meta.url)))
+
 export default defineConfig({
   resolve: {
     alias: {
       chibivue: path.resolve(dirname, '../../packages'),
     },
   },
+  plugins: [chibivue()],
 })
